requirements.txt
APScheduler == 3.6.3
SQLAlchemy == 1.3.18
discord.py == 1.3.4
python_dotenv == 0.14.0
pytz == 2020.1
requests == 2.18.4
setuptools == 39.0.1
psutil~=5.7.2